+++ hyp3lib/get_orb.py
@@ -6,6 +6,7 @@ import os
 import re
 import sys
 from datetime import datetime
+from typing import Optional, Tuple
 
 import requests
 from lxml import html
@@ -16,7 +17,7 @@ from urllib3.util.retry import Retry
 from hyp3lib import OrbitDownloadError
 from hyp3lib.fetch import download_file
 
-ESA_AUTH = ('gnssguest', 'gnssguest')
+ESA_AUTH_TOKEN_URL = 'https://identity.dataspace.copernicus.eu/auth/realms/CDSE/protocol/openid-connect/token'
 
 
 def _get_asf_orbit_url(orbit_type, platform, timestamp):
@@ -57,28 +58,40 @@ def _get_asf_orbit_url(orbit_type, platform, timestamp):
 
 
 def _get_esa_orbit_url(orbit_type: str, platform: str, start_time: datetime, end_time: datetime):
-    search_url = 'https://scihub.copernicus.eu/gnss/api/stub/products.json'
+    search_url = 'https://catalogue.dataspace.copernicus.eu/odata/v1/Products'
 
     date_format = '%Y-%m-%dT%H:%M:%SZ'
     params = {
-        'filter': f'(platformname:Sentinel-1 AND producttype:{orbit_type} AND filename:{platform}* '
-                  f'AND beginPosition:[* TO {start_time.strftime(date_format)}] '
-                  f'AND endPosition:[{end_time.strftime(date_format)} TO NOW])',
-        'limit': 1,
-        'offset': 0,
-        'sortedby': 'ingestiondate',
-        'order': 'desc',
+        '$filter': f"Collection/Name eq 'SENTINEL-1' and "
+                   f"startswith(Name, '{platform}_OPER_{orbit_type}_OPOD_') and "
+                   f"ContentDate/Start lt {start_time.strftime(date_format)} and "
+                   f"ContentDate/End gt {end_time.strftime(date_format)}",
+        '$orderby': 'Name desc',
+        '$top': 1,
     }
 
-    response = requests.get(search_url, params=params, auth=ESA_AUTH)
+    response = requests.get(search_url, params=params)
     response.raise_for_status()
     data = response.json()
 
     orbit_url = None
-    if data['products']:
-        uuid = data['products'][0]['uuid']
-        orbit_url = f"https://scihub.copernicus.eu/gnss/odata/v1/Products('{uuid}')/$value"
+    if data['value']:
+        product_id = data['value'][0]['Id']
+        orbit_url = f'https://zipper.dataspace.copernicus.eu/download/{product_id}'
+
     return orbit_url
+
+
+def _get_esa_auth_token(username: str, password: str) -> str:
+    data = {
+        'client_id': 'cdse-public',
+        'grant_type': 'password',
+        'username': username,
+        'password': password,
+    }
+    response = requests.post(ESA_AUTH_TOKEN_URL, data=data)
+    response.raise_for_status()
+    return response.json()['access_token']
 
 
 def get_orbit_url(granule: str, orbit_type: str = 'AUX_POEORB', provider: str = 'ESA'):
@@ -108,7 +121,8 @@ def get_orbit_url(granule: str, orbit_type: str = 'AUX_POEORB', provider: str = 
 
 
 def downloadSentinelOrbitFile(
-        granule: str, directory: str = '', providers=('ESA', 'ASF'), orbit_types=('AUX_POEORB', 'AUX_RESORB')
+        granule: str, directory: str = '', providers=('ESA', 'ASF'), orbit_types=('AUX_POEORB', 'AUX_RESORB'),
+        esa_credentials: Optional[Tuple[str, str]] = None,
 ):
     """Download a Sentinel-1 Orbit file
 
@@ -117,24 +131,24 @@ def downloadSentinelOrbitFile(
         directory: Directory to save the orbit files into
         providers: Iterable of providers to attempt to download the orbit file from, in order of preference
         orbit_types: Iterable of orbit file types to attempt to download, in order of preference
+        esa_credentials: Copernicus Data Space Ecosystem (CDSE) username and password
 
     Returns: Tuple of:
         orbit_file: The downloaded orbit file
         provider: The provider used to download the orbit file from
 
     """
-    provider_auth_map = {
-        'ESA': ESA_AUTH,
-        'ASF': None,  # use netrc instead of HTTP Basic Auth
-    }
+    if 'ESA' in providers and esa_credentials is None:
+        raise ValueError('esa_credentials must be provided if ESA in providers')
     for orbit_type in orbit_types:
         for provider in providers:
             try:
                 url = get_orbit_url(granule, orbit_type, provider=provider)
+                token = _get_esa_auth_token(*esa_credentials) if provider == 'ESA' else None
                 orbit_file = download_file(
                     url,
                     directory=directory,
-                    auth=provider_auth_map[provider]
+                    token=token,
                 )
                 if orbit_file:
                     return orbit_file, provider