+++ tests/test_get_orb.py
@@ -9,12 +9,44 @@ _GRANULE = 'S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8'
 
 
 @responses.activate
-def test_download_sentinel_orbit_file_esa(tmp_path):
-    responses.add(responses.GET, 'https://foo.bar/hello.txt', body='content')
+def test_get_esa_auth_token():
+    url = get_orb.ESA_AUTH_TOKEN_URL
+    request_payload = {
+        'client_id': 'cdse-public',
+        'grant_type': 'password',
+        'username': 'myUsername',
+        'password': 'myPassword',
+    }
+    response_payload = {
+        'access_token': 'ABC123',
+    }
+    responses.add(
+        responses.POST,
+        url=url,
+        match=[responses.matchers.urlencoded_params_matcher(request_payload)],
+        json=response_payload,
+    )
 
-    with patch('hyp3lib.get_orb.get_orbit_url', return_value='https://foo.bar/hello.txt'):
-        responses.add(responses.GET, 'https://foo.bar/hello.txt', body='content')
-        orbit_file, provider = get_orb.downloadSentinelOrbitFile(_GRANULE, providers=('ESA',), directory=str(tmp_path))
+    assert get_orb._get_esa_auth_token('myUsername', 'myPassword') == 'ABC123'
+
+
+@responses.activate
+def test_download_sentinel_orbit_file_esa(tmp_path):
+    responses.add(
+        method=responses.GET,
+        url='https://foo.bar/hello.txt',
+        body='content',
+        match=[responses.matchers.header_matcher({'Authorization': 'Bearer test-token'})],
+    )
+
+    with patch('hyp3lib.get_orb.get_orbit_url', return_value='https://foo.bar/hello.txt'), \
+            patch('hyp3lib.get_orb._get_esa_auth_token', return_value='test-token'):
+        orbit_file, provider = get_orb.downloadSentinelOrbitFile(
+            _GRANULE,
+            providers=('ESA',),
+            directory=str(tmp_path),
+            esa_credentials=('user', 'pass'),
+        )
 
     assert provider == 'ESA'
     assert os.path.exists(orbit_file)
@@ -23,30 +55,32 @@ def test_download_sentinel_orbit_file_esa(tmp_path):
 
 @responses.activate
 def test_get_orbit_url_esa_poeorb():
-    search_url = 'https://scihub.copernicus.eu/gnss/api/stub/products.json' \
-                 '?filter=%28platformname%3ASentinel-1+AND+producttype%3AAUX_POEORB+AND+filename%3AS1A%2A+' \
-                 'AND+beginPosition%3A%5B%2A+TO+2015-06-21T12%3A02%3A20Z%5D+' \
-                 'AND+endPosition%3A%5B2015-06-21T12%3A02%3A32Z+TO+NOW%5D%29' \
-                 '&limit=1&offset=0&sortedby=ingestiondate&order=desc'
-    search_response = {'products': [{'uuid': 'myUUID'}]}
+    search_url = ('https://catalogue.dataspace.copernicus.eu/odata/v1/Products?'
+                  '%24filter=Collection%2FName+eq+%27SENTINEL-1%27+and+'
+                  'startswith%28Name%2C+%27S1A_OPER_AUX_POEORB_OPOD_%27%29+and+'
+                  'ContentDate%2FStart+lt+2015-06-21T12%3A02%3A20Z+and+'
+                  'ContentDate%2FEnd+gt+2015-06-21T12%3A02%3A32Z&'
+                  '%24orderby=Name+desc&%24top=1')
+    search_response = {'value': [{'Id': 'myProductId'}]}
     responses.add(responses.GET, search_url, json=search_response)
 
     orbit_url = get_orb.get_orbit_url(_GRANULE, provider='ESA')
-    assert orbit_url == "https://scihub.copernicus.eu/gnss/odata/v1/Products('myUUID')/$value"
+    assert orbit_url == 'https://zipper.dataspace.copernicus.eu/download/myProductId'
 
 
 @responses.activate
 def test_get_orbit_url_esa_resorb():
-    search_url = 'https://scihub.copernicus.eu/gnss/api/stub/products.json' \
-                 '?filter=%28platformname%3ASentinel-1+AND+producttype%3AAUX_RESORB+AND+filename%3AS1A%2A+' \
-                 'AND+beginPosition%3A%5B%2A+TO+2015-06-21T12%3A02%3A20Z%5D+' \
-                 'AND+endPosition%3A%5B2015-06-21T12%3A02%3A32Z+TO+NOW%5D%29' \
-                 '&limit=1&offset=0&sortedby=ingestiondate&order=desc'
-    search_response = {'products': [{'uuid': 'myUUID'}]}
+    search_url = ('https://catalogue.dataspace.copernicus.eu/odata/v1/Products?'
+                  '%24filter=Collection%2FName+eq+%27SENTINEL-1%27+and+'
+                  'startswith%28Name%2C+%27S1A_OPER_AUX_RESORB_OPOD_%27%29+and+'
+                  'ContentDate%2FStart+lt+2015-06-21T12%3A02%3A20Z+and+'
+                  'ContentDate%2FEnd+gt+2015-06-21T12%3A02%3A32Z&'
+                  '%24orderby=Name+desc&%24top=1')
+    search_response = {'value': [{'Id': 'myProductId'}]}
     responses.add(responses.GET, search_url, json=search_response)
 
     orbit_url = get_orb.get_orbit_url(_GRANULE, provider='ESA', orbit_type='AUX_RESORB')
-    assert orbit_url == "https://scihub.copernicus.eu/gnss/odata/v1/Products('myUUID')/$value"
+    assert orbit_url == 'https://zipper.dataspace.copernicus.eu/download/myProductId'
 
 
 def test_get_orbit_url_asf():